requirements.txt
gradio
huggingface_hub
sentence-transformers
faiss-cpu
numpy
PyPDF2>=3.0.0

